Build fix: common: eventfd only exists on Linux
The project fails to build on non-Linux platforms since eventfd is a
Linux-exclusive facility.

From the project's point of view, the eventfd util belongs in libcommon.
However, we will need to write a wrapper if it ends up being used on
non-Linux platform.

For the moment, exclude it from the internal libcommon sources when
building on non-Linux platforms as it is only used on the session
daemon.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: If528ac522d72a95e87a4f496531ea679a81030a2