Coverity warning: sessiond: uncaught exception in main
Coverity reports:
1508778 Uncaught exception
If the exception is ever thrown, the program will crash.

In main: A C++ exception is thrown but never caught (CWE-248)

In particular, Coverity reports that pthread_mutex_lock can error-out,
which would cause an lttng::posix_error exception to be thrown.

This isn't particularly likely to happen, but it is nonetheless
preferable to catch exceptions at the top-level and log them before
exiting.

Change-Id: I4f7a52db3c9cd19764e7d12fd62afa60af99dabd
Reported-by: Coverity Scan
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>