Tests: enable test_callstack on 32-bit x86
The test application used by test_callstack defines syscall
stubs (instead of using the libc) for both 32 and 64 bit variants of
x86. This limits the architectures that can run this test.

However, the `uname` check is erroneous as it assumes that `uname -m`
will return x86 for both architectures. It doesn't; it returns i386 or
i686 (and I presume others) for 32-bit kernels.

As such, the check is modified to check for '86' which, as far as I
know, doesn't clash with other architectures.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I267a81fadb2f1db6b135c6886ae5700184f2ae5b